move jetcolormap from mvs to image
