+++ src/aliceVision/image/jetColorMap.cpp
