+++ src/aliceVision/image/jetColorMap.hpp
